requirements.txt
# Flask & server
Flask>=3.0.0

# LLM and LangGraph/LangChain/OpenAI
langchain>=0.2.0
langgraph>=0.1.0
langchain-openai>=0.1.0
openai>=1.50.0
python-dotenv>=1.0.1

# Document processing
docxtpl>=0.16.7
python-docx>=1.1.0
pydantic>=2.0
beautifulsoup4>=4.1.0
lxml
arabic-reshaper
python-bidi
# Utilities
typing_extensions>=4.12.2
tika
uvicorn>=0.30.0